large_star: A40 137->128 (same (u16)-on-s16 signedness bug as B90); B90 flat at 45, four lever families exhausted
Categorical rule: a pin can only ADD a constraint — it can never
synthesize the extra live range that produced ref's numbering. "Ref has
one more value" residues are outside the entire pin/clobber family.

+++ wip/large_star_stone_block.c
@@ -409,7 +409,7 @@ void sub_08120A40(struct Unknown82 *x)
     y12 = yu;
 
     j = 0;
-    for (i = 0; i < x->unkB4; i++) {
+    for (i = 0; i < (u16)x->unkB4; i++) {
             u16 dx = gUnk_08357D30[x->unkB6 * 0x12 + j];
             u16 dy;
             u8 ex, ey;
